fix: use f32s

+++ src/math/util.rs
@@ -1,9 +1,9 @@
 /// Radians to degrees
-pub fn rad_to_deg(rad: f64) -> f64 {
-    rad * (180.0 / std::f64::consts::PI)
+pub fn rad_to_deg(rad: f32) -> f32 {
+    rad * (180.0 / std::f32::consts::PI)
 }
 
 // Degrees to radians
-pub fn deg_ro_rad(deg: f64) -> f64 {
-    deg * (std::f64::consts::PI / 180.0)
+pub fn deg_ro_rad(deg: f32) -> f32 {
+    deg * (std::f32::consts::PI / 180.0)
 }
